capstone_project_flowers/requirements.txt
tensorflow>=2.6.0
streamlit>=1.0.0
Pillow>=8.0.0
numpy>=1.19.2
matplotlib>=3.3.2 